Update suspense loading
