+++ src/App.tsx
@@ -5,8 +5,7 @@ import Cast from 'pages/Cast'
 import CreateBurnerWallet from 'pages/CreateBurnerWallet'
 import Footer from 'components/Footer'
 import Landing from 'pages/Landing'
-import Logo from 'icons/Logo'
-import Navbar from 'components/navbar/Navbar'
+import Navbar from 'components/navbar'
 import Privacy from 'pages/Privacy'
 import ProtectedRoute from 'components/ui/ProtectedRoute'
 import ScrollToTop from 'components/ui/ScrollToTop'
@@ -19,7 +18,6 @@ import classnames, {
   minHeight,
   width,
 } from 'classnames/tailwind'
-import useAccount from 'hooks/useAccount'
 
 const pageContainer = classnames(
   display('flex'),
@@ -32,35 +30,12 @@ const bodyContainer = classnames(
   margin('mx-4', 'mb-auto', 'body:mx-auto')
 )
 
-function NavBarSuspended() {
-  const { account } = useAccount()
-
-  return (
-    <Navbar
-      logo={<Logo />}
-      account={account}
-      needNetworkChange={false}
-      logoText="SealCaster"
-    />
-  )
-}
-
 export default function () {
   return (
     <Router>
       <ScrollToTop>
         <div className={pageContainer}>
-          <Suspense
-            fallback={
-              <Navbar
-                logo={<Logo />}
-                needNetworkChange={false}
-                logoText="SealCaster"
-              />
-            }
-          >
-            <NavBarSuspended />
-          </Suspense>
+          <Navbar />
           <div className={bodyContainer}>
             <Route path="/">
               <Landing />
